refactor(websocket): extract JobExecutionManager from the runner
T3 of docs/websocket-runner-refactor-plan.md. UnifiedWebSocketRunner was one
class doing six jobs, and `activeJobs` was mutated from three of them. The job
region — admission gates, cost estimation, the concurrency queue, run start-up,
the message stream, reconnect, cancel and terminal-status persistence — now
lives in session/job-execution.ts behind the ClientSession seam T2 introduced.
It has one owner, and it can be read and changed apart from chat and RPC.

Behavior is unchanged: same frames in, same messages out, same queue and abort
semantics. The synchronous slot reservation and the slotCounters leak
accounting move with their doc comments attached.

Chat, RPC and the command switch still reach the moved members through thin
delegators on the runner, so T4 and T5 — extracted in parallel from the same
base — merge on top of this without touching their own call sites. The
delegators go away in T7/T8.
